backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
scikit-learn==1.4.0
pandas==2.2.0
joblib==1.3.2
supabase==2.3.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
pydantic==2.6.0
